refactor agent to use base agent interface
